Fix #3: QueryRead: Use mixin instead of clone for Dijit compatibility
Dijit's _SearchMixin (for ComboBox and FilteringSelect) includes some
funky code involving setting the query to a RegExp to take advantage
of dojo/store/Memory, but with a toString function to attempt to
interoperate with other stores.

Unfortunately, QueryRead's default use of lang.clone ends up creating
a new RegExp, discarding the helpful toString method.

The simple solution is to switch to using mixin instead which does not
attempt special logic on various object types. Technically this comes
at the expense of support for nested structures, but that should be
beyond the scope of what would be expected in a query object anyway.